split code into reusable functions and container
