+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "memory",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
     "react": "^16.7.0",
