+++ src/App.js
@@ -1,12 +1,12 @@
 import React, { Component } from 'react';
 import './App.css';
-import ScreenComponent from '../src/components/ScreenComponent';
+import ScreenContainer from '../src/components/ScreenContainer';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <ScreenComponent/>
+        <ScreenContainer/>
       </div>
     );
   }
