+++ src/common/Footer.js
@@ -1,0 +1,13 @@
+import React from 'react';
+
+class Footer extends React.Component{
+
+    render(){
+        return(
+            <div></div>
+        )
+    }
+
+}
+
+export default Footer;
